+++ src/c/cat.c
@@ -1,16 +1,27 @@
 int main(int argc, char *argv[]) {
     char buffer[512];
-    if (argc != 2) {
-        die("Usage: cat <filename>\n");
+    int fd;
+    int needs_close = 0;
+    if (argc == 1) {
+        fd = STDIN;
+    } else if (argc == 2) {
+        fd = open(argv[1], O_RDONLY);
+        if (fd < 0) {
+            die("File not found\n");
+        }
+        needs_close = 1;
+    } else {
+        die("Usage: cat [filename]\n");
     }
-    int fd = open(argv[1], O_RDONLY);
-    if (fd < 0) {
-        die("File not found\n");
-    }
-    int bytes;
-    do {
-        bytes = read(fd, buffer, 512);
+    while (1) {
+        int bytes = read(fd, buffer, 512);
+        if (bytes <= 0) {
+            break;
+        }
         write(STDOUT, buffer, bytes);
-    } while (bytes > 0);
-    close(fd);
+    }
+    if (needs_close) {
+        close(fd);
+    }
+    return 0;
 }